liveness constraints: draw the rest of the owl
